% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tof_tbl.R
\name{as_flowFrame}
\alias{as_flowFrame}
\alias{as_flowFrame.tof_tbl}
\title{Coerce an object into a \code{\link[flowCore]{flowFrame}}}
\usage{
as_flowFrame(x, ...)

\method{as_flowFrame}{tof_tbl}(x, ...)
}
\arguments{
\item{x}{A tof_tbl.}

\item{...}{Unused.}
}
\value{
A \code{\link[flowCore]{flowFrame}}

A \code{\link[flowCore]{flowFrame}}. Note that all non-numeric
columns in `x` will be removed.
}
\description{
Coerce an object into a \code{\link[flowCore]{flowFrame}}

Coerce a tof_tbl into a \code{\link[flowCore]{flowFrame}}
}
\examples{
NULL

NULL
}
