% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tof_tbl.R
\name{as_tof_tbl.flowSet}
\alias{as_tof_tbl.flowSet}
\title{Convert an object into a tof_tbl}
\usage{
\method{as_tof_tbl}{flowSet}(flow_data, sep = "|")
}
\arguments{
\item{flow_data}{A FlowSet}

\item{sep}{A string to use to separate the antigen name and its associated
metal in the column names of the output tibble. Defaults to "|".}
}
\value{
a `tof_tbl`
}
\description{
Convert an object into a tof_tbl
}
