% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cosine_similarity}
\alias{cosine_similarity}
\title{Find the cosine similarity between two vectors}
\usage{
cosine_similarity(x, y)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector}
}
\value{
a scalar value representing the cosine similarity between x and y
}
\description{
Find the cosine similarity between two vectors
}
