% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dot}
\alias{dot}
\title{Find the dot product between two vectors.}
\usage{
dot(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}
}
\value{
The dot product between x and y.
}
\description{
Find the dot product between two vectors.
}
\examples{
NULL

}
