% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_extension}
\alias{get_extension}
\title{Find the extension for a file}
\usage{
get_extension(filename)
}
\arguments{
\item{filename}{A string representing the name of a file in its local directory}
}
\value{
The the file extension of `filename`
}
\description{
Find the extension for a file
}
