% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowFrame_methods.R
\name{group_by.flowFrame}
\alias{group_by.flowFrame}
\title{Group a flowFrame into a flowSet using one or more variables.}
\usage{
\method{group_by}{flowFrame}(.data, ..., .add = FALSE, .drop = dplyr::group_by_drop_default(.data))
}
\arguments{
\item{.data}{A \code{\link[flowCore]{flowFrame}}}

\item{...}{Unquoted variables or columns to group by according to .data's
\code{\link[flowCore]{featureNames}}.}

\item{.add}{Unused.}

\item{.drop}{Unused.}
}
\value{
A \code{\link[flowCore]{flowSet}} containing one
\code{\link[flowFrame]{flowFrame}} for each of the unique combinations of columns
selected in .... Metadata about grouping columns will be stored in the output
\code{\link[flowCore]{flowSet}}'s \code{\link[flowCore]{pData}}.
}
\description{
Group a flowFrame into a flowSet using one or more variables.
}
