% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{l2_normalize}
\alias{l2_normalize}
\title{L2 normalize an input vector x to a length of 1}
\usage{
l2_normalize(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a vector of length length(x) with a magnitude of 1
}
\description{
L2 normalize an input vector x to a length of 1
}
