% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{magnitude}
\alias{magnitude}
\title{Find the magnitude of a vector.}
\usage{
magnitude(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A scalar value (the magnitude of x).
}
\description{
Find the magnitude of a vector.
}
