% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tof_tbl.R
\name{make_flowcore_annotated_data_frame}
\alias{make_flowcore_annotated_data_frame}
\title{Make the AnnotatedDataFrame needed for the flowFrame class}
\usage{
make_flowcore_annotated_data_frame(maxes_and_mins)
}
\arguments{
\item{maxes_and_mins}{a data.frame containing information about the max
and min values of each channel to be saved in the flowFrame.}
}
\value{
An AnnotatedDataFrame.
}
\description{
Make the AnnotatedDataFrame needed for the flowFrame class
}
\examples{
NULL
}
