% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metal_masterlist}
\alias{metal_masterlist}
\title{A character vector of metal name patterns supported by tidytof.}
\format{
A character vector in which each entry is a pattern that tidytof searches
for in every CyTOF channel in input .fcs files. These patterns are an amalgamate
of example .fcs files sampled from the studies linked below.
}
\source{
\url{https://github.com/kara-davis-lab/DDPR}
\url{https://cytobank.org/nolanlab/reports/Levine2015.html}
\url{https://cytobank.org/nolanlab/reports/Spitzer2015.html}
\url{https://cytobank.org/nolanlab/reports/Spitzer2017.html}
\url{https://community.cytobank.org/cytobank/projects/609}
}
\usage{
data(metal_masterlist)
}
\description{
A character vector used by `tof_read_fcs` and `tof_read_data` to detect and
parse which CyTOF metals correspond to each channel in an input .fcs file.
}
\keyword{datasets}
