% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_helpers.R
\name{new_tof_model}
\alias{new_tof_model}
\title{Constructor for a tof_model.}
\usage{
new_tof_model(model, recipe)
}
\arguments{
\item{model}{A glmnet model.}

\item{recipe}{A prepped recipe object.}
}
\value{
A `tof_model`, an S3 class that includes a trained glmnet model and
the recipe used to perform its associated preprocessing.
}
\description{
Constructor for a tof_model.
}
