% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidytof_example_data}
\alias{tidytof_example_data}
\title{Get paths to tidytof example data}
\usage{
tidytof_example_data(dataset_name = NULL)
}
\arguments{
\item{dataset_name}{Name of the dataset you want to access. If NULL,
the names of the datasets (each of which is from a different study)
will be listed.}
}
\value{
A character vector of file paths where the requested .fcs
files are located. If `dataset_name` is NULL, a character vector of
dataset names (that can be used as values for `dataset_name`) is
returned instead.
}
\description{
tidytof comes bundled with a number of sample .fcs files in its
inst/extdata directory. This function makes them easy to access.
}
\examples{
tidytof_example_data()
tidytof_example_data(dataset_name = "phenograph")

}
