% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_helpers.R
\name{tof_assess_model_new_data}
\alias{tof_assess_model_new_data}
\title{Compute a trained elastic net model's performance metrics using new_data.}
\usage{
tof_assess_model_new_data(tof_model, new_data)
}
\arguments{
\item{tof_model}{A `tof_model` trained using \code{\link{tof_train_model}}}

\item{new_data}{A tibble of new observations that should be used to evaluate
the `tof_model`'s performance.}
}
\value{
A list of performance metrics whose components depend on the model type.
}
\description{
Compute a trained elastic net model's performance metrics using new_data.
}
