% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_helpers.R
\name{tof_assess_model_tuning}
\alias{tof_assess_model_tuning}
\title{Access a trained elastic net model's performance metrics using its tuning data.}
\usage{
tof_assess_model_tuning(tof_model)
}
\arguments{
\item{tof_model}{A `tof_model` trained using \code{\link{tof_train_model}}}
}
\value{
A list of performance metrics whose components depend on the model type.
}
\description{
Access a trained elastic net model's performance metrics using its tuning data.
}
