% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_helpers.R
\name{tof_clean_metric_names}
\alias{tof_clean_metric_names}
\title{Title}
\usage{
tof_clean_metric_names(metric_tibble, model_type)
}
\arguments{
\item{metric_tibble}{TO DO}

\item{model_type}{TO DO}
}
\value{
TO DO
}
\description{
TO DO
}
