% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tof_compute_km_curve}
\alias{tof_compute_km_curve}
\title{Compute a Kaplan-Meier curve from sample-level survival data}
\usage{
tof_compute_km_curve(survival_curves)
}
\arguments{
\item{survival_curves}{A tibble from which the Kaplan-Meier curve will be
computed. Each row must represent an observation and must have two
columns named "time_to_event" and "event".}
}
\value{
A tibble with 3 columns: time_to_event, survival_probability, and
is_censored (whether or not an event was censored at that timepoint).
}
\description{
Compute a Kaplan-Meier curve from sample-level survival data
}
