% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/developmental_classifier_helpers.R
\name{tof_cosine_dist}
\alias{tof_cosine_dist}
\title{A function for finding the cosine distance between each of the rows of a numeric
matrix and a numeric vector.}
\usage{
tof_cosine_dist(matrix, vector)
}
\arguments{
\item{matrix}{A numeric matrix.}

\item{vector}{A numeric vector.}
}
\value{
A numeric vector of distances of length `nrow(matrix)` in which the
ith entry represents the cosine distance between the ith row of `matrix` and
`vector`.
}
\description{
A function for finding the cosine distance between each of the rows of a numeric
matrix and a numeric vector.
}
\examples{
NULL
}
