% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality_reduction.R
\name{tof_dr}
\alias{tof_dr}
\title{Apply dimensionality reduction to a CyTOF dataset}
\usage{
tof_dr(tof_tibble, method = c("pca", "tsne", "umap"), ...)
}
\arguments{
\item{tof_tibble}{A `tof_tibble`.}

\item{method}{A method of dimensionality reduction. Currently, PCA, tSNE, and
UMAP embedding are supported.}

\item{...}{Arguments to be passed to the tof_dr_* function corresponding to
the embedding method. See \code{\link{tof_dr_pca}}, \code{\link{tof_dr_tsne}}, and
\code{\link{tof_dr_umap}}.}
}
\value{
A tibble with the same number of rows as `tof_tibble`, each representing
a single cell. Each of the `num_comp` columns represents each cell's embedding
in the calculated embedding space.
}
\description{
Apply dimensionality reduction to a CyTOF dataset
}
\examples{
NULL

}
