% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality_reduction.R
\name{tof_dr_pca}
\alias{tof_dr_pca}
\title{Perform principal component analysis on CyTOF data}
\usage{
tof_dr_pca(
  tof_tibble,
  pca_vars = where(tof_is_numeric),
  num_comp = 5,
  threshold = NA,
  center = TRUE,
  scale = TRUE
)
}
\arguments{
\item{tof_tibble}{A `tof_tibble`}

\item{pca_vars}{Unquoted column names indicating which columns in `tof_tibble` to
use in computing the principal components. Defaults to all numeric columns
in `tof_tibble`. Supports tidyselect helpers.}

\item{num_comp}{The number of PCA components to calculate. Defaults
to 5. See \code{\link[recipes]{step_pca}}.}

\item{threshold}{A double between 0 and 1 representing the fraction of total
variance that should be covered by the components returned in the output. See
\code{\link[recipes]{step_pca}}.}

\item{center}{A boolean value indicating if each column should be centered to
mean 0 before PCA analysis. Defaults to TRUE.}

\item{scale}{A boolean value indicating if each column should be scaled to
standard deviation = 1 before PCA analysis. Defaults to TRUE.}
}
\value{
A tibble with the same number of rows as `tof_tibble`, each representing
a single cell. Each of the `num_comp` columns represents each cell's embedding
in the calculated principal component space.
}
\description{
This function calculates principal components using single-cell data from a `tof_tibble`.
}
\examples{
NULL

}
