% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{tof_find_density}
\alias{tof_find_density}
\title{Find the kernel density estimate for a list of cells}
\usage{
tof_find_density(my_cells, ...)
}
\arguments{
\item{my_cells}{A numeric (double or integer) vector representing single-cell measurements.}

\item{...}{}
}
\value{

}
\description{
Find the kernel density estimate for a list of cells
}
\examples{
NULL
}
