% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tof_find_emd}
\alias{tof_find_emd}
\title{Find the earth-mover's distance between two numeric vectors}
\usage{
tof_find_emd(vec_1, vec_2, num_bins = 100)
}
\arguments{
\item{vec_1}{A numeric vector.}

\item{vec_2}{A numeric vector.}

\item{num_bins}{An integer number of bins to use when performing kernel
density estimation on the two vectors. Defaults to 100.}
}
\value{
A double (of length 1) representing the EMD between the two vectors.
}
\description{
Find the earth-mover's distance between two numeric vectors
}
