% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tof_find_knn_rann}
\alias{tof_find_knn_rann}
\title{Find the k-nearest neighbors of each cell in a high-dimensional cytometry dataset.}
\usage{
tof_find_knn_rann(
  .data,
  k = min(10, nrow(.data)),
  distance_function = c("euclidean", "cosine"),
  ...
)
}
\arguments{
\item{.data}{A `tof_tibble` or `tibble` in which each row represents a cell
and each column represents a high-dimensional cytometry measurement.}

\item{k}{An integer indicating the number of nearest neighbors to return for
each cell.}

\item{distance_function}{A string indicating which distance function to use
for the nearest-neighbor calculation. Options include "euclidean"
(the default) and "cosine" distances.}

\item{...}{Optional additional arguments to pass to RANN::nn2}
}
\value{
A list with two elements: "neighbor_ids" and "neighbor_distances,"
both of which are n by k matrices (in which n is the number of cells in the
input `.data`. The [i,j]-th entry of "neighbor_ids" represents the row index
for the j-th nearest neighbor of the cell in the i-th row of `.data`.
The [i,j]-th entry of "neighbor_distances" represents the distance between
those two cells according to `distance_function`.
}
\description{
Find the k-nearest neighbors of each cell in a high-dimensional cytometry dataset.
}
\examples{
sim_data <-
    dplyr::tibble(
        cd45 = rnorm(n = 1000),
        cd38 = rnorm(n = 1000),
        cd34 = rnorm(n = 1000),
        cd19 = rnorm(n = 1000)
    )

# Find the 10 nearest neighbors of each cell in the dataset
tof_find_knn(
    .data = sim_data,
    k = 10,
    distance_function = "euclidean"
)

# Find the 10 approximate nearest neighbors
tof_find_knn(
    .data = sim_data,
    k = 10,
    distance_function = "euclidean"
)

}
