% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{tof_generate_palette}
\alias{tof_generate_palette}
\title{Generate a color palette using tidytof.}
\usage{
tof_generate_palette(num_colors)
}
\arguments{
\item{num_colors}{An integer specifying the number of colors you'd like to generate.}
}
\value{
A character vector of hex codes specifying the colors in the palette.
}
\description{
This function generates a color palette based on the color palette of the
author's favorite pokemon.
}
\examples{
tof_generate_palette(num_colors = 5L)

}
