% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_helpers.R
\name{tof_get_model_mixture}
\alias{tof_get_model_mixture}
\title{Get a `tof_model`'s optimal mixture (alpha) value}
\usage{
tof_get_model_mixture(tof_model)
}
\arguments{
\item{tof_model}{A tof_model}
}
\value{
A numeric value
}
\description{
Get a `tof_model`'s optimal mixture (alpha) value
}
