% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_helpers.R
\name{tof_get_model_outcomes}
\alias{tof_get_model_outcomes}
\title{Get a `tof_model`'s outcome variable name(s)}
\usage{
tof_get_model_outcomes(tof_model)
}
\arguments{
\item{tof_model}{A tof_model}
}
\value{
A character vector
}
\description{
Get a `tof_model`'s outcome variable name(s)
}
