% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_helpers.R
\name{tof_get_model_penalty}
\alias{tof_get_model_penalty}
\title{Get a `tof_model`'s optimal penalty (lambda) value}
\usage{
tof_get_model_penalty(tof_model)
}
\arguments{
\item{tof_model}{A tof_model}
}
\value{
A numeric value
}
\description{
Get a `tof_model`'s optimal penalty (lambda) value
}
