% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_helpers.R
\name{tof_get_model_training_data}
\alias{tof_get_model_training_data}
\title{Get a `tof_model`'s training data}
\usage{
tof_get_model_training_data(tof_model)
}
\arguments{
\item{tof_model}{A tof_model}
}
\value{
A tibble of (non-preprocessed) training data used to fit the model
}
\description{
Get a `tof_model`'s training data
}
