% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_helpers.R
\name{tof_get_model_type}
\alias{tof_get_model_type}
\title{Get a `tof_model`'s model type}
\usage{
tof_get_model_type(tof_model)
}
\arguments{
\item{tof_model}{A tof_model}
}
\value{
A string
}
\description{
Get a `tof_model`'s model type
}
