% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_helpers.R
\name{tof_get_model_x}
\alias{tof_get_model_x}
\title{Get a `tof_model`'s processed predictor matrix (for glmnet)}
\usage{
tof_get_model_x(tof_model)
}
\arguments{
\item{tof_model}{A tof_model}
}
\value{
An x value formatted for glmnet
}
\description{
Get a `tof_model`'s processed predictor matrix (for glmnet)
}
