% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_helpers.R
\name{tof_get_model_y}
\alias{tof_get_model_y}
\title{Get a `tof_model`'s processed outcome variable matrix (for glmnet)}
\usage{
tof_get_model_y(tof_model)
}
\arguments{
\item{tof_model}{A tof_model}
}
\value{
A y value formatted for glmnet
}
\description{
Get a `tof_model`'s processed outcome variable matrix (for glmnet)
}
