% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{tof_histogram}
\alias{tof_histogram}
\title{Generate univariate histograms of CyTOF data from a `tof_tibble`.}
\usage{
tof_histogram(
  tof_tibble = NULL,
  channel_var = NULL,
  group_var = NULL,
  ordered = FALSE,
  color_option = "plasma",
  lower_quantile = 0.01,
  upper_quantile = 0.99,
  ...
)
}
\arguments{
\item{tof_tibble}{A tibble, data.frame, or something that can be coerced into either}

\item{channel_var}{An unquoted variable name indicating which channel should be plotted}

\item{group_var}{An unquoted variable name indicating which variable should be used to
break the cells into different histograms along the overlay.}

\item{ordered}{A logical indicating the the histograms along the y axis should
be ordered by decreasing median channel value.}

\item{color_option}{Color palette name to be passed to scale_color_viridis as argument `option`.
Choices include "inferno", "plasma", "viridis", "magma", and "cividis".}

\item{lower_quantile}{lowest quantile in `channel_var` to include in the density plot. All cells with
`channel_var` values lower than this will be filtered out of the plot.}

\item{upper_quantile}{highest quantile in `channel_var` to include in the density plot. All cells with
`channel_var` values higher than this will be filtered out of the plot.}

\item{...}{Optional additional arguments to be passed to `geom_tof_ridges()`}
}
\value{
A ggplot object.
}
\description{
Plots a series of overlaid histograms for a given CyTOF channel broken up
by a given grouping variable.
}
\examples{
NULL

}
