% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tof_is_numeric}
\alias{tof_is_numeric}
\title{Find if a vector is numeric}
\usage{
tof_is_numeric(.vec)
}
\arguments{
\item{.vec}{A vector.}
}
\value{
A boolean value indicating if .vec is of type integer or double.
}
\description{
This function takes an input vector `.vec` and checks if it is either an
integer or a double (i.e. is the type of vector that might encode CyTOF
measurements).
}
