% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tof_knn_density_legacy}
\alias{tof_knn_density_legacy}
\title{Find the KNN density estimate for each cell in a CyTOF dataset (legacy version).}
\usage{
tof_knn_density_legacy(
  neighbor_ids,
  neighbor_distances,
  method = c("mean_distance", "sum_distance", "spade"),
  normalize = TRUE
)
}
\arguments{
\item{neighbor_ids}{A n by k matrix returned by `tof_find_knn` representing the row indices
of the k nearest neighbors of each of the n cells in a CyTOF dataset.}

\item{neighbor_distances}{A n by k matrix returned by `tof_find_knn` representing the pairwise distances
between a cell and each of its k nearest neighbors in a CyTOF dataset.}

\item{method}{A string indicating how the relative density for each cell should be
calculated from the distances between it and each of its k nearest neighbors. Options are
"mean_distance" (the default; estimates the relative density for a cell's neighborhood by
taking the negative average of the distances to its nearest neighbors) and "sum_distance"
(estimates the relative density for a cell's neighborhood by taking the negative sum of the
distances to its nearest neighbors).}

\item{normalize}{TO DO}
}
\value{
a vector of length N (number of cells) with the ith
}
\description{
Find the KNN density estimate for each cell in a CyTOF dataset (legacy version).
}
