% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{tof_plot_all_histograms}
\alias{tof_plot_all_histograms}
\title{Title}
\usage{
tof_plot_all_histograms(
  tof_tibble,
  group_var,
  out_path = NULL,
  label_prefix = "",
  label_suffix = "",
  width,
  height,
  device = "pdf",
  ...
)
}
\arguments{
\item{tof_tibble}{a tibble, data.frame, or somehing that can be coerced into either
of these}

\item{group_var}{an unquoted variable name indicating which variable should be used to
break the cells into different histograms.}

\item{out_path}{file path to the directory in which the plots should be saved}

\item{label_prefix}{An optional string to concatenate to the beginning of each file name.}

\item{label_suffix}{An optional string to concatenate to the beginning of each file name.}

\item{width}{The width for the plots being saved.}

\item{height}{The height of the plots being saved.}

\item{device}{"jpg", "tif", or "pdf" file format}

\item{...}{Optional additional arguments to pass to `tof_histogram()`}
}
\value{
Save all density plots to the file path specified by `out_path`.
}
\description{
Plots a series of overlain histograms for all CyTOF channels in your dataset broken up
by a given grouping variable.
}
\examples{
NULL

}
