% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{tof_plot_model}
\alias{tof_plot_model}
\title{Plot the results of a glmnet model fit on sample-level data.}
\usage{
tof_plot_model(tof_model, new_data, theme = ggplot2::theme_bw())
}
\arguments{
\item{tof_model}{A `tof_model` trained using \code{\link{tof_train_model}}}

\item{new_data}{A tibble of new observations for which a plot should be made.
If new_data isn't provided, the plot will be made using the training data used to
fit the model. Alternatively, the string "tuning_data" can be provided, and the
plot will be generated using the predictions generated during model tuning.}

\item{theme}{A ggplot2 theme to apply to the plot
Defaults to \code{\link[ggplot2]{theme_bw}}}
}
\value{
A ggplot object. If the `tof_model` is a linear model, a scatterplot
of the predicted outcome vs. the true outcome will be returned. If the `tof_model`
is a two-class model, an ROC curve will be returned. If the `tof_model` is a
multiclass model, a one-versus-all ROC curve will be returned for each class.
If `tof_model` is a survival model, a Kaplan-Meier curve will be returned.
}
\description{
Plot the results of a glmnet model fit on sample-level data.
}
