% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{tof_plot_model_survival}
\alias{tof_plot_model_survival}
\title{Plot the results of a survival glmnet model fit on sample-level data.}
\usage{
tof_plot_model_survival(
  tof_model,
  new_data,
  censor_size = 2.5,
  theme = ggplot2::theme_bw()
)
}
\arguments{
\item{tof_model}{A `tof_model` trained using \code{\link{tof_train_model}}}

\item{new_data}{A tibble of new observations for which a plot should be made.
If new_data isn't provided, the plot will be made using the training data used to
fit the model. Alternatively, the string "tuning_data" can be provided, and the
plot will be generated using the predictions generated during model tuning.}

\item{censor_size}{A numeric value indicating how large to plot the tick marks
representing censored values in the Kaplan-Meier curve.}

\item{theme}{A ggplot2 theme to apply to the plot.
Defaults to \code{\link[ggplot2]{theme_bw}}}
}
\value{
A ggplot object. Specifically, a Kaplan-Meier curve.
}
\description{
Plot the results of a survival glmnet model fit on sample-level data.
}
