% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{tof_save_figures}
\alias{tof_save_figures}
\title{Save each entry in a list-column of ggplot figures in a tibble.}
\usage{
tof_save_figures(
  plot_tibble = NULL,
  out_path = NULL,
  label_column = NULL,
  plot_column = plots,
  label_prefix = "",
  label_suffix = "",
  width,
  height,
  device = "pdf",
  ...
)
}
\arguments{
\item{plot_tibble}{A tibble containing a column of ggplot objects}

\item{out_path}{File path to the directory in which the plots should be saved}

\item{label_column}{An unquoted column name representing a column in `plot_tibble`
containing strings that can be used to uniquely identify each plot in `plot_tibble`.}

\item{plot_column}{An unquoted column name indicating which column in `plot_tibble`
contains the plots to be saved.}

\item{label_suffix}{Optional string to concatenate to the end of each file name.}

\item{width}{The width for the plots being saved.}

\item{height}{The height for the plots being saved.}

\item{device}{"jpg", "tiff", or "pdf" file format.}

\item{...}{Additional arguments to pass to `ggsave()`}

\item{label_prefixan}{Optional string to concatenate to the beginning of each file name.}
}
\description{
Saves a list of plots contained in a plot_tibble (useful for nested tibbles).
}
\examples{
NULL
}
