% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient-level_modeling.R
\name{tof_train_classifier}
\alias{tof_train_classifier}
\title{Title}
\usage{
tof_train_classifier(
  feature_tibble,
  response_col,
  predictor_cols = -{     {         response_col     } },
  split_method = c("k-fold", "simple", "bootstrap"),
  ...,
  model_type = c("two-class", "multiclass"),
  normalize_predictors = TRUE,
  remove_zv_predictors = TRUE,
  impute_missing_predictors = FALSE,
  grid_type = c("entropy", "regular"),
  grid_size = 100,
  run_parallel = FALSE
)
}
\arguments{
\item{feature_tibble}{TO DO}

\item{response_col}{TO DO}

\item{predictor_cols}{TO DO}

\item{split_method}{TO DO}

\item{...}{TO DO}

\item{model_type}{TO DO}

\item{normalize_predictors}{TO DO}

\item{remove_zv_predictors}{TO DO}

\item{impute_missing_predictors}{TO DO}

\item{grid_type}{TO DO}

\item{grid_size}{TO DO}

\item{run_parallel}{TO DO}
}
\value{
TO DO
}
\description{
Title
}
