% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient-level_modeling.R
\name{tof_train_survival}
\alias{tof_train_survival}
\title{Title}
\usage{
tof_train_survival(
  feature_tibble,
  time_col,
  event_col,
  predictor_cols = -c({     {         time_col     } }, {     {         event_col     }
    }),
  split_method = c("k-fold", "simple", "bootstrap"),
  ...,
  remove_zv_predictors = TRUE,
  impute_missing_predictors = FALSE,
  grid_type = c("entropy", "regular"),
  grid_size = 100,
  best_model_type = c("best", "best with sparsity"),
  run_parallel = FALSE
)
}
\arguments{
\item{feature_tibble}{TO DO}

\item{time_col}{TO DO}

\item{event_col}{TO DO}

\item{predictor_cols}{TO DO}

\item{split_method}{TO DO}

\item{...}{TO DO}

\item{remove_zv_predictors}{TO DO}

\item{impute_missing_predictors}{TO DO}

\item{grid_type}{TO DO}

\item{grid_size}{TO DO}

\item{best_model_type}{TO DO}

\item{run_parallel}{TO DO}
}
\value{
TO DO
}
\description{
Title
}
