% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_helpers.R
\name{tof_tune_glmnet}
\alias{tof_tune_glmnet}
\title{Title}
\usage{
tof_tune_glmnet(
  split_data,
  prepped_recipe,
  hyperparameter_grid,
  model_type,
  outcome_cols,
  optimization_metric = "tidytof_default",
  num_cores = 1
)
}
\arguments{
\item{split_data}{TO DO}

\item{prepped_recipe}{TO DO}

\item{hyperparameter_grid}{TO DO}

\item{model_type}{TO DO}

\item{outcome_cols}{TO DO}

\item{optimization_metric}{TO DO}

\item{num_cores}{TO DO}
}
\value{
TO DO
}
\description{
TO DO
}
