% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{tof_write_fcs}
\alias{tof_write_fcs}
\title{Write a series of .fcs files from a tof_tbl}
\usage{
tof_write_fcs(tof_tibble, group_cols, out_path, sep = "_", file_name)
}
\arguments{
\item{tof_tibble}{A `tof_tbl` or a `tibble`.}

\item{group_cols}{Unquoted names of the columns in `tof_tibble` that should
be used to group cells into separate files. Supports tidyselect helpers. Defaults to
NULL (all cells are written into a single file).}

\item{out_path}{A system path indicating the directory where the output .csv
files should be saved. If the directory doesn't exist, it will be created.}

\item{sep}{Delimiter that should be used between each of the values of `group_cols`
to create the output .fcs file names. Defaults to "_".}

\item{file_name}{If `group_cols` isn't specified, the name (without an extension)
that should be used for the saved .csv file.}
}
\value{
This function does not return anything. Instead, it has the side-effect
of saving .fcs files to `out_path`.
}
\description{
This function takes a given `tof_tbl` and writes the single-cell data
it contains into .fcs files within the directory located at `out_path`. The
`group_cols` argument specifies how the rows of the `tof_tbl` (each cell)
should be broken into separate .fcs files
}
\examples{
NULL

}
\seealso{
Other input/output functions:
\code{\link[=tof_read_data]{tof_read_data()}},
\code{\link[=tof_write_csv]{tof_write_csv()}},
\code{\link[=tof_write_data]{tof_write_data()}}
}
\concept{input/output functions}
