% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{where}
\alias{where}
\title{Select variables with a function}
\usage{
where(fn)
}
\arguments{
\item{fn}{A function that returns TRUE or FALSE (technically, a predicate function).
Can also be a purrr-like formula.}
}
\value{
A predicate that can be used to select columns from a data.frame.
}
\description{
This is a copy of \code{\link[tidyselect]{where}}, a selection helper that
selects the variables for which a predicate function returns TRUE. See
 \code{\link[tidyselect]{language}} for more details about tidyselection.
}
\details{
This help file was replicated verbatim from \code{\link[tidyselect]{tidyselect-package}}.
}
\examples{
NULL

}
\references{
Lionel Henry and Hadley Wickham (2021). tidyselect:
Select from a Set of Strings. R package version 1.1.1.
https://CRAN.R-project.org/package=tidyselect
}
