% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tof_tbl.R
\name{as_tof_tbl}
\alias{as_tof_tbl}
\title{Coerce flowFrames or flowSets into tof_tbl's.}
\usage{
as_tof_tbl(flow_data, sep = "|")
}
\arguments{
\item{flow_data}{A flowFrame or flowSet}

\item{sep}{A string indicating which symbol should be used to separate
antigen names and metal names in the columns of the output tof_tbl.}
}
\value{
A tof_tbl.
}
\description{
Coerce flowFrames or flowSets into tof_tbl's.
}
\examples{
input_file <- dir(tidytof_example_data("aml"), full.names = TRUE)[[1]]

input_flowframe <- flowCore::read.FCS(input_file)

tof_tibble <- as_tof_tbl(input_flowframe)

}
