% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{new_tof_tibble}
\alias{new_tof_tibble}
\title{Create a new tof_tibble.}
\usage{
new_tof_tibble(x = tibble::tibble(), panel = tibble::tibble())
}
\arguments{
\item{x}{}

\item{panel}{}
}
\value{

}
\description{
Create a new tof_tibble.
}
\examples{

}
