% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tof_tbl.R
\name{new_tof_tibble}
\alias{new_tof_tibble}
\title{Constructor for tof_tibble.}
\usage{
new_tof_tibble(x = tibble::tibble(), panel = tibble::tibble())
}
\arguments{
\item{x}{}

\item{panel}{}
}
\value{
A `tof_tbl`, an tibble extension that tracks a few other attributes
that are useful for CyTOF data analysis.
}
\description{
Constructor for tof_tibble.
}
\examples{

}
