% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tof_tbl.R
\name{new_tof_tibble}
\alias{new_tof_tibble}
\title{Constructor for a tof_tibble.}
\usage{
new_tof_tibble(x = tibble::tibble(), panel = tibble::tibble())
}
\arguments{
\item{x}{A data.frame or tibble containing single-cell mass cytometry data
such that rows are cells and columns are CyTOF measurements.}

\item{panel}{A data.frame or tibble containing information about the panel
for the mass cytometry data in x.}
}
\value{
A `tof_tbl`, an tibble extension that tracks a few other attributes
that are useful for CyTOF data analysis.
}
\description{
Constructor for a tof_tibble.
}
