% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient-level_modeling.R
\name{tof_assess_model}
\alias{tof_assess_model}
\title{TO DO}
\usage{
tof_assess_model(tof_model, new_data = NULL, ...)
}
\arguments{
\item{tof_model}{TO DO}

\item{new_data}{TO DO}

\item{...}{TO DO}
}
\value{
TO DO
}
\description{
TO DO
}
