% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{tof_cluster}
\alias{tof_cluster}
\title{Cluster CyTOF data.}
\usage{
tof_cluster(tof_tibble, method, ..., add_col = TRUE)
}
\arguments{
\item{tof_tibble}{A `tof_tbl` or `tibble`.}

\item{method}{A string indicating which clustering methods should be used. Valid
values include "flowsom", "phenograph", "kmeans", "ddpr", and "xshift".}

\item{...}{Additional arguments to pass onto the `tof_cluster_*`
function family member corresponding to the chosen method.}

\item{add_col}{A boolean value indicating if the output should column-bind the
cluster ids of each cell as a new column in `tof_tibble` (TRUE, the default) or if
a single-column tibble including only the cluster ids should be returned (FALSE).}
}
\value{
A `tof_tbl` or `tibble` If add_col = FALSE, it will have a single column encoding
the cluster ids for each cell in `tof_tibble`. If add_col = TRUE, it will have
ncol(tof_tibble) + 1 columns: each of the (unaltered) columns in `tof_tibble`
plus an additional column encoding the cluster ids.
}
\description{
This function is a wrapper around {tidytof}'s tof_cluster_* function family.
It performs clustering on CyTOF data using a user-specified method (of 5 choices)
and each method's corresponding input parameters
}
\seealso{
Other clustering functions: 
\code{\link{tof_cluster_ddpr}()},
\code{\link{tof_cluster_flowsom}()},
\code{\link{tof_cluster_kmeans}()},
\code{\link{tof_cluster_phenograph}()}
}
\concept{clustering functions}
