% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{tof_cluster_ddpr}
\alias{tof_cluster_ddpr}
\title{Title}
\usage{
tof_cluster_ddpr(
  healthy_tibble,
  cancer_tibble,
  healthy_cell_labels,
  cluster_vars = where(tof_is_numeric),
  distance_function = c("mahalanobis", "cosine", "pearson"),
  num_cores = 1,
  parallel_vars = NULL,
  return_distances = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{healthy_tibble}{A `tibble` or `tof_tibble` containing cells from only
healthy control samples (i.e. not disease samples).}

\item{cancer_tibble}{}

\item{healthy_cell_labels}{A character or integer vector of length `nrow(healthy_tibble)`.
Each entry in this vector should represent the cell subpopulation label (or cluster id) for
the corresponding row in `healthy_tibble`.}

\item{cluster_vars}{}

\item{distance_function}{}

\item{num_cores}{}

\item{return_distances}{}

\item{verbose}{}

\item{...}{}
}
\value{

}
\description{
Title
}
\examples{
NULL

}
