% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient-level_modeling.R
\name{tof_create_grid}
\alias{tof_create_grid}
\title{Create a {glmnet} hyperparameter search grid of a specified size}
\usage{
tof_create_grid(
  penalty_values,
  mixture_values,
  grid_type = c("regular", "entropy"),
  grid_size = 25
)
}
\arguments{
\item{penalty_values}{A numeric vector of all penalty values to include in the
hyperparamter grid. If unspecified, {tidytof}'s default values will be used:
10^(-10), 10^(-7.5), 10^(-5), 10^(-2.5), 10^(0).}

\item{mixture_values}{A numeric vector of all penalty values to include in the
hyperparamter grid. If unspecified, {tidytof}'s default values will be used:
`seq(0, 1, length.out = 10)`.}

\item{grid_type}{A string indicating if a regular grid ("regular"; the default) or maximum-entropy
("entropy") search grid should be returned. Ignored if both `penalty_values` and
`mixture_values` are provided. If "entropy" is chosen, the {dials} package
is required.}

\item{grid_size}{An integer indicating how many combinations of penalty and mixture
values should be searched during model tuning. Defaults to 25 (5 unique values
of `penalty` and 5 unique values of `mixture`).}
}
\value{
A tibble with two numeric columns: `penalty` and `mixture`.
}
\description{
This function creates a hyperparameter search grid (in the form of a
\code{\link[tibble]{tibble}}) specifying the search space for the two
hyperparameters of a generalized linear model using the {glmnet} package:
the regularization penalty term
and the lasso/ridge regression mixture term. Either a regular or maximum-entropy
search grid (of a specified size) or a grid with a user-specified range of
hyperparameter values can be created.
}
\examples{
tof_create_grid()

}
