% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsampling.R
\name{tof_downsample_constant}
\alias{tof_downsample_constant}
\title{Title}
\usage{
tof_downsample_constant(tof_tibble, group_cols, num_cells)
}
\arguments{
\item{tof_tibble}{TO DO}

\item{group_cols}{TO DO}

\item{num_cells}{TO DO}
}
\value{
TO DO
}
\description{
Title
}
