% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsampling.R
\name{tof_downsample_density}
\alias{tof_downsample_density}
\title{Title}
\usage{
tof_downsample_density(
  tof_tibble,
  group_cols,
  density_cols = where(tof_is_numeric),
  outlier_percentile = 0.01,
  target_percentile = 0.1,
  num_neighbors = 15,
  knn_distance_function = c("euclidean", "cosine"),
  ...
)
}
\arguments{
\item{tof_tibble}{A `tibble` or `tof_tibble`.}

\item{group_cols}{TO DO}

\item{density_cols}{TO DO}

\item{outlier_percentile}{TO DO}

\item{target_percentile}{TO DO}

\item{num_neighbors}{TO DO}

\item{knn_distance_function}{TO DO}

\item{...}{TO DO}
}
\value{
TO DO
}
\description{
Title
}
