% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsampling.R
\name{tof_downsample_prop}
\alias{tof_downsample_prop}
\title{Title}
\usage{
tof_downsample_prop(tof_tibble, group_cols, prop_cells)
}
\arguments{
\item{tof_tibble}{TO DO}

\item{group_cols}{TO DO}

\item{prop_cells}{TO DO}
}
\value{
TO DO
}
\description{
Title
}
