% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{tof_extract_central_tendency}
\alias{tof_extract_central_tendency}
\title{Extract the central tendencies of CyTOF markers in each cluster in a `tof_tibble`.}
\usage{
tof_extract_central_tendency(
  tof_tibble,
  cluster_col,
  group_cols = NULL,
  marker_cols = where(tof_is_numeric),
  central_tendency_function = stats::median,
  format = c("wide", "long")
)
}
\arguments{
\item{tof_tibble}{A `tof_tibble` or a `tibble` in which each row represents a
single cell and each column represents a CyTOF measurement or a piece of metadata
(i.e. cluster id, patient id, etc.) about each cell.}

\item{cluster_col}{An unquoted column name indicating which column in `tof_tibble`
stores the cluster ids of the cluster to which each cell belongs.
Cluster labels can be produced via any method the user chooses - including manual gating,
any of the functions in the `tof_cluster_*` function family, or any other method.}

\item{group_cols}{Unquoted column names representing which columns in `tof_tibble`
should be used to break the rows of `tof_tibble` into subgroups for the feature
extraction calculation. Defaults to NULL (i.e. performing the extraction without subgroups).}

\item{marker_cols}{Unquoted column names representing which columns in `tof_tibble`
(i.e. which CyTOF protein measurements) should be included in the feature extraction
calculation. Defaults to all numeric (integer or double) columns.
Supports tidyselection.}

\item{central_tendency_function}{The function that will be used to calculate
the measurement of central tendency for each cluster (to be used
as the dependent variable in the linear model). Defaults to \code{\link[stats]{median}}.}

\item{format}{A string indicating if the data should be returned in "wide" format
(the default; each cluster feature is given its own column) or in "long" format
(each cluster feature is provided as its own row).}
}
\value{
A tibble.

If format == "wide", the tibble will have 1 row for each combination of
the grouping variables provided in `group_cols` and one column for each grouping variable,
one column for each extracted feature (the central tendency of a given marker in a given cluster).
The names of each column containing cluster features is obtained using the following pattern:
"\{marker_id\}@\{cluster_id\}".

If format == "long", the tibble will have 1 row for each combination of the grouping variables
in `group_cols`, each cluster id (i.e. level) in `cluster_col`, and each marker in `marker_cols`.
It will have one column for each grouping variable, one column for the cluster ids, one
column for the CyTOF channel names, and one column (`value`) containing the features.
}
\description{
This feature extraction function calculates a user-specified measurement of central tendency
(i.e. median or mode) of the cells in each cluster in a `tof_tibble` across a
user-specified selection of CyTOF markers. These calculations can be done either
overall (across all cells in the dataset) or after breaking down the cells into
subgroups using `group_cols`.
}
