% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{tof_extract_emd}
\alias{tof_extract_emd}
\title{Title}
\usage{
tof_extract_emd(
  tof_tibble,
  cluster_col,
  group_cols = NULL,
  marker_cols = where(tof_is_numeric),
  stimulation_col,
  basal_level,
  format = c("wide", "long"),
  num_bins = 100
)
}
\arguments{
\item{tof_tibble}{TO DO}

\item{cluster_col}{TO DO}

\item{group_cols}{TO DO}

\item{marker_cols}{TO DO}

\item{stimulation_col}{TO DO}

\item{basal_level}{TO DO}

\item{format}{TO DO}

\item{num_bins}{TO DO}
}
\value{
TO DO
}
\description{
Title
}
