% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{tof_extract_features}
\alias{tof_extract_features}
\title{Title}
\usage{
tof_extract_features(
  tof_tibble,
  cluster_col,
  group_cols = NULL,
  stimulation_col = NULL,
  lineage_cols,
  signaling_cols,
  central_tendency_function = mean,
  signaling_method = c("threshold", "emd", "jsd", "central tendency"),
  basal_level = NULL,
  ...
)
}
\arguments{
\item{tof_tibble}{TO DO}

\item{cluster_col}{TO DO}

\item{group_cols}{TO DO}

\item{stimulation_col}{TO DO}

\item{lineage_cols}{TO DO}

\item{signaling_cols}{TO DO}

\item{central_tendency_function}{TO DO}

\item{signaling_method}{TO DO}

\item{basal_level}{TO DO}

\item{...}{TO DO}
}
\value{
TO DO
}
\description{
Title
}
