% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_helpers.R
\name{tof_find_best}
\alias{tof_find_best}
\title{Find the optimal hyperparameters for an elastic net model from candidate performance metrics}
\usage{
tof_find_best(performance_metrics, model_type, optimization_metric)
}
\arguments{
\item{performance_metrics}{TO DO}

\item{model_type}{TO DO}

\item{optimization_metric}{TO DO}
}
\value{
TO DO
}
\description{
Find the optimal hyperparameters for an elastic net model from candidate performance metrics
}
