% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{tof_plot_cells_embedding}
\alias{tof_plot_cells_embedding}
\title{Title}
\usage{
tof_plot_cells_embedding(
  tof_tibble,
  embedding_cols,
  color_col,
  facet_cols,
  embedding_method = c("pca", "tsne", "umap"),
  point_alpha = 1,
  theme = ggplot2::theme_bw(),
  ...
)
}
\arguments{
\item{tof_tibble}{TO DO}

\item{embedding_cols}{TO DO}

\item{color_col}{TO DO}

\item{facet_cols}{TO DO}

\item{embedding_method}{TO DO}

\item{point_alpha}{TO DO}

\item{theme}{TO DO}

\item{...}{TO DO}
}
\value{
TO DO
}
\description{
description
}
\seealso{
Other visualization functions: 
\code{\link{tof_plot_cells_layout}()}
}
\concept{visualization functions}
