% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{tof_plot_cells_layout}
\alias{tof_plot_cells_layout}
\title{Title}
\usage{
tof_plot_cells_layout(
  tof_tibble,
  knn_cols = where(tof_is_numeric),
  color_col,
  facet_cols,
  num_neighbors = 5,
  graph_type = c("weighted", "unweighted"),
  graph_layout = "fr",
  distance_function = c("euclidean", "cosine"),
  knn_error = 0,
  edge_alpha = 0.25,
  node_size = 2,
  theme = ggplot2::theme_void(),
  ...
)
}
\arguments{
\item{tof_tibble}{TO DO}

\item{knn_cols}{TO DO}

\item{color_col}{TO DO}

\item{facet_cols}{TO DO}

\item{num_neighbors}{TO DO}

\item{graph_type}{TO DO}

\item{graph_layout}{TO DO}

\item{distance_function}{TO DO}

\item{knn_error}{TO DO}

\item{edge_alpha}{TO DO}

\item{node_size}{TO DO}

\item{theme}{TO DO}

\item{...}{TO DO}
}
\value{
TO DO
}
\description{
description
}
