% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{tof_plot_sc_dr}
\alias{tof_plot_sc_dr}
\title{Title}
\usage{
tof_plot_sc_dr(
  tof_tibble,
  dr_cols,
  color_col,
  facet_cols,
  dr_method = c("pca", "tsne", "umap"),
  point_alpha = 1,
  theme = ggplot2::theme_bw(),
  ...
)
}
\arguments{
\item{tof_tibble}{}

\item{dr_cols}{}

\item{color_col}{}

\item{facet_cols}{}

\item{dr_method}{}

\item{point_alpha}{}

\item{theme}{}

\item{...}{}
}
\value{

}
\description{
description
}
\examples{
NULL

}
