% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{tof_plot_sc_dr}
\alias{tof_plot_sc_dr}
\title{Title}
\usage{
tof_plot_sc_dr(
  tof_tibble,
  dr_cols,
  color_col,
  facet_cols,
  dr_method = c("pca", "tsne", "umap"),
  point_alpha = 1,
  theme = ggplot2::theme_bw(),
  ...
)
}
\arguments{
\item{tof_tibble}{TO DO}

\item{dr_cols}{TO DO}

\item{color_col}{TO DO}

\item{facet_cols}{TO DO}

\item{dr_method}{TO DO}

\item{point_alpha}{TO DO}

\item{theme}{TO DO}

\item{...}{TO DO}
}
\value{
TO DO
}
\description{
description
}
