% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{tof_plot_sc_layout}
\alias{tof_plot_sc_layout}
\title{Title}
\usage{
tof_plot_sc_layout(
  tof_tibble,
  knn_cols = where(tof_is_numeric),
  color_col,
  facet_cols,
  num_neighbors = 5,
  graph_type = c("weighted", "unweighted"),
  graph_layout = "fr",
  distance_function = c("euclidean", "cosine"),
  knn_error = 0,
  edge_alpha = 0.25,
  node_size = 2,
  theme = ggplot2::theme_void(),
  ...
)
}
\arguments{
\item{tof_tibble}{}

\item{knn_cols}{}

\item{color_col}{}

\item{facet_cols}{}

\item{num_neighbors}{}

\item{graph_type}{}

\item{graph_layout}{}

\item{distance_function}{}

\item{knn_error}{}

\item{edge_alpha}{}

\item{node_size}{}

\item{theme}{}

\item{...}{}
}
\value{

}
\description{
description
}
\examples{
NULL


}
