% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{tof_postprocess}
\alias{tof_postprocess}
\title{Preprocess transformed CyTOF data.}
\usage{
tof_postprocess(
  tof_tibble = NULL,
  metadata_vars = NULL,
  channel_vars = where(is.numeric),
  redo_noise = FALSE,
  transform_fun = function(x) rev_asinh(x, shift_factor = 0, scale_factor = 0.2)
)
}
\arguments{
\item{tof_tibble}{A `tof_tibble` or a `tibble`.}

\item{metadata_vars}{Unquoted column names of columns in `tof_tibble` that should
not be computed over, i.e. file names, patient names, stimulation names, etc.
Supports tidy selection using tidy_select helpers.
Not currently used.}

\item{channel_vars}{A vector of non-quoted variables representing columns that contain
single-cell protein measurements. Anything that works in the first argument of
dplyr::across will work. See ?across. Supports tidy selection using tidyselect
helpers. If nothing is specified, the default is to transform all numeric columns.}

\item{redo_noise}{A boolean value indicating whether to add  uniform noise that
to each CyTOF measurement for aesthetic and visualization purposes. See \href{https://pubmed.ncbi.nlm.nih.gov/30277658/}{this paper}.
Defaults to FALSE}

\item{transform_fun}{A vectorized function to apply to each column specified by
`channel_vars` for post-processing. Defaults to \code{\link{rev_asinh}} transformation
(with a cofactor of 5).}
}
\value{
A `tof_tibble` with identical dimensions to the input `tof_tibble`, with all
columns specified in channel_vars transformed using `transform_fun` (with noise
added or not removed depending on `redo_noise`).
}
\description{
This function transforms a `tof_tibble` of transformed ion counts from a mass
cytometer back into something that looks more like an .fcs file that Fluidigm
software generates.
}
