% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient-level_modeling.R
\name{tof_predict}
\alias{tof_predict}
\title{TO DO}
\usage{
tof_predict(
  tof_model,
  new_data,
  prediction_type = c("response", "class", "link"),
  ...
)
}
\arguments{
\item{tof_model}{TO DO}

\item{new_data}{TO DO}

\item{prediction_type}{TO DO}

\item{...}{TO DO}
}
\value{
TO DO
}
\description{
TO DO
}
