% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tof_io.R
\name{tof_read_fcs}
\alias{tof_read_fcs}
\title{Read data from an .fcs file or a directory of .fcs files.}
\usage{
tof_read_fcs(file_path = NULL)
}
\arguments{
\item{file_path}{A file path. It can lead to either a single .fcs file or
to a directory filled with .fcs files.}
}
\value{
An [m by n+1] tibble in which each row represents a single cell (of m
total in the dataset) and each column represents a metal measurement
(of n total in the dataset). In addition, the last column of the tibble will
represent the filename from which the cell was read.
}
\description{
Read data from an .fcs file or a directory of .fcs files.
}
\examples{
NULL
}
