% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient-level_modeling.R
\name{tof_train_model}
\alias{tof_train_model}
\title{TO DO}
\usage{
tof_train_model(
  split_data,
  predictor_cols,
  response_col = NULL,
  time_col = NULL,
  event_col = NULL,
  model_type = c("linear", "two-class", "multiclass", "survival"),
  hyperparameter_grid = tof_create_grid(),
  standardize_predictors = TRUE,
  remove_zv_predictors = FALSE,
  impute_missing_predictors = FALSE,
  optimization_metric = "tidytof_default",
  best_model_type = c("best", "best with sparsity"),
  num_cores = 1
)
}
\arguments{
\item{split_data}{TO DO}

\item{predictor_cols}{TO DO}

\item{response_col}{TO DO}

\item{time_col}{TO DO}

\item{event_col}{TO DO}

\item{model_type}{TO DO}

\item{hyperparameter_grid}{TO DO}

\item{standardize_predictors}{TO DO}

\item{remove_zv_predictors}{TO DO}

\item{impute_missing_predictors}{TO DO}

\item{optimization_metric}{TO DO}

\item{best_model_type}{TO DO}

\item{num_cores}{TO DO}
}
\value{
TO DO
}
\description{
TO DO
}
