% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tof_io.R
\name{tof_write_out}
\alias{tof_write_out}
\title{Write data to a file}
\usage{
tof_write_out(
  tof_tibble = NULL,
  group_vars = "file_names",
  out_folder_path = NULL,
  format = NULL
)
}
\arguments{
\item{tof_tibble}{a tibble, data.frame, or something that can be coerced into either}

\item{group_vars}{variables that should be used to group cells into individual files.
The default is to group cells based on the column "file_names" (as produced by tof_read_fcs).}

\item{out_folder_path}{file path to the folder where the output files should be saved}

\item{format}{format for the files being written. Currently supports .csv and .fcs files}
}
\value{

}
\description{
Write data (in the form of a tibble) into either a .csv or an .fcs file for storage.
}
